requirements.txt
pytest~=8.3.4
selenium~=4.27.1
requests~=2.31.0
pillow~=11.0.0